Inline Control class code into the damn class
Not sure why this was split out into InstanceMethods and ClassMethods but it's really difficult to follow over just having the methods for the class defined in it.
